+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/presentation/ReceiveScreen.kt
@@ -1,6 +1,5 @@
 package com.liftley.sync360.features.sync.presentation
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -12,24 +11,16 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.weight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.automirrored.filled.Article
 import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.ChevronRight
-import androidx.compose.material.icons.filled.Description
 import androidx.compose.material.icons.filled.Inbox
-import androidx.compose.material.icons.filled.PermMedia
-import androidx.compose.material.icons.filled.Schedule
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CenterAlignedTopAppBar
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
@@ -43,29 +34,25 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import coil3.compose.AsyncImage
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.liftley.sync360.features.sync.domain.model.PendingIncomingOffer
 import com.liftley.sync360.features.sync.domain.model.TransferDirection
-import com.liftley.sync360.features.sync.domain.model.TransferFilePreview
-import com.liftley.sync360.features.sync.presentation.components.FileTransferErrorCard
-import com.liftley.sync360.features.sync.presentation.components.FileTransferProgressCard
+import com.liftley.sync360.features.sync.presentation.components.transfer.FileTransferErrorCard
+import com.liftley.sync360.features.sync.presentation.components.transfer.FileTransferProgressCard
 import com.liftley.sync360.features.sync.presentation.components.Sync360Surface
 import com.liftley.sync360.features.sync.presentation.components.SyncBottomNavigationBar
 import com.liftley.sync360.features.sync.presentation.components.formatBytes
+import com.liftley.sync360.features.sync.presentation.navigation.SyncRoute
 import com.liftley.sync360.features.sync.presentation.receive.ItemTransferStatus
 import com.liftley.sync360.features.sync.presentation.receive.ProgressFileRow
 import com.liftley.sync360.features.sync.presentation.receive.ReceiveProposalCard
 import com.liftley.sync360.features.sync.presentation.receive.ReceivedFileRow
 import com.liftley.sync360.features.sync.presentation.receive.getFileStatuses
-import com.liftley.sync360.features.sync.presentation.navigation.SyncRoute
 import kotlinx.coroutines.flow.Flow
 
 @OptIn(ExperimentalMaterial3Api::class)
